Fake News detection
Fake news detection has been in order to detect the fake news that spread and this helps to avoid incorrect information being sent or spread in public. this kind of model helps famous personalities to save themselves from fake news that spreads about them in public.

this model is built using a passive-aggressive classifier from sklearn. 

this model has 87.8% accuracy